+++ CHANGELOG.md
@@ -21,6 +21,7 @@
 - `XIT FINCH`: Added an "Equity Growth" chart
 - `XIT FLOW`: Shows daily production, consumption, and value of each material across bases
 - `XIT PLNT`: Per-planet settings for bases you own
+- `XIT SET`: Added financial data merging tool in `XIT SET FIN`
 - `XIT STO`: Per-base storage analysis
 - Added a notification count badge to the browser tab favicon (`favicon-notification-badge`)
 - Added a right-click market context menu for materials (`market-context-menu`)